requirements.txt
azure-search-documents>=11.4.0
azure-identity==1.15.0
python-dotenv==1.0.0
fastapi==0.115.12
uvicorn==0.24.0
pydantic==2.11.5
pydantic-settings>=2.4.0
azure-storage-blob>=12.19.0
azure-ai-formrecognizer==3.3.0
azure-ai-vision-imageanalysis==1.0.0b1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
torch==2.1.1
transformers==4.35.2

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.1  # For testing FastAPI endpoints

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.18.1
webdriver-manager==4.0.1

# LangChain
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
playwright==1.42.0  # Required for some LangChain web scrapers

# Added from the code block
pydantic>=2.0.0

# New Azure SDK packages
azure-core>=1.29.5 